PR response, consolidate base class and move base collection classes
