+++ EasyReflectometry/sample/__init__.py
@@ -3,11 +3,11 @@ from .assemblies.gradient_layer import GradientLayer
 from .assemblies.multilayer import Multilayer
 from .assemblies.repeating_multilayer import RepeatingMultilayer
 from .assemblies.surfactant_layer import SurfactantLayer
-from .elements.layer_collection import LayerCollection
 from .elements.layers.layer import Layer
 from .elements.layers.layer_area_per_molecule import LayerAreaPerMolecule
-from .elements.material_collection import MaterialCollection
+from .elements.layers.layer_collection import LayerCollection
 from .elements.materials.material import Material
+from .elements.materials.material_collection import MaterialCollection
 from .elements.materials.material_mixture import MaterialMixture
 from .elements.materials.material_solvated import MaterialSolvated
 from .sample import Sample
