+++ EasyReflectometry/sample/assemblies/multilayer.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
-from ..elements.layer_collection import LayerCollection
 from ..elements.layers.layer import Layer
+from ..elements.layers.layer_collection import LayerCollection
 from .base_assembly import BaseAssembly
