+++ EasyReflectometry/sample/elements/layers/layer_collection.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 __author__ = 'github.com/arm61'
 
-from .base_element_collection import BaseElementCollection
-from .layers.layer import Layer
+from ..base_element_collection import BaseElementCollection
+from .layer import Layer
 
 
 class LayerCollection(BaseElementCollection):
