+++ tests/sample/assemblies/test_repeating_multilayer.py
@@ -13,8 +13,8 @@ from numpy.testing import assert_raises
 
 from EasyReflectometry.calculators import CalculatorFactory
 from EasyReflectometry.sample.assemblies.repeating_multilayer import RepeatingMultilayer
-from EasyReflectometry.sample.elements.layer_collection import LayerCollection
 from EasyReflectometry.sample.elements.layers.layer import Layer
+from EasyReflectometry.sample.elements.layers.layer_collection import LayerCollection
 from EasyReflectometry.sample.elements.materials.material import Material
